display parameters

+++ app/src/main/java/com/example/phonelocalization/Particle.kt
@@ -1,6 +1,6 @@
 package com.example.phonelocalization
 
-class Particle(x: Int, y : Int, weight : Int) {
+class Particle(x: Double, y : Double, weight : Int) {
     var x = x
     var y = y
     var weight = weight
